[강운덕] [LUCYSUS-1744] spanId가 long으로 변경됨에 따라 annotationMapper에서 int값이 long으로 변경안된 부분이 있어 수정함.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-commons/trunk@2749 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/common/bo/AnnotationBo.java
@@ -15,7 +15,7 @@ public class AnnotationBo {
     private static final int VERSION_SIZE = 1;
 
     private byte version = 0;
-    private int spanId;
+    private long spanId;
 
     private int key;
 
@@ -36,11 +36,11 @@ public class AnnotationBo {
         this.byteValue = transcoder.encode(value, this.valueType);
     }
 
-    public int getSpanId() {
+    public long getSpanId() {
         return spanId;
     }
 
-    public void setSpanId(int spanId) {
+    public void setSpanId(long spanId) {
         this.spanId = spanId;
     }
